refactor: split useTechnician
