+++ src/components/editTechnicianForm/editTechnicianForm.tsx
@@ -1,4 +1,3 @@
-import useTechnician from "@/hook/useTechnician";
 import editingTechnicianSchema, {
   editingTechnicianType,
 } from "@/schemas/editingTechnician";
@@ -8,6 +7,7 @@ import { useEffect } from "react";
 import { useForm, FormProvider } from "react-hook-form";
 import ButtonPrimary from "../buttonPrimary/buttonPrimary";
 import FormField from "../formField/formField";
+import useTechnician from "@/hook/useTechnician/useTechnician";
 
 type EditTechnicianFormProps = {
   technician: Technician;
