+++ src/hook/useTechnician/types.ts
@@ -1,0 +1,12 @@
+export type GetAllTechnicianParams = {
+    pageParam: number;
+    searchTerm?: SearchTerm;
+  };
+
+export type SearchTerm = {
+    name?: string;
+    email?: string;
+    phone?: number;
+}
+
+export type SearchField = keyof SearchTerm;
